+++ ComposeUI/Sources/ComposeUI/Components/InnerShadowLayer.swift
@@ -140,7 +140,9 @@ open class InnerShadowLayer: CALayer {
     }
 
     if let animationTiming {
-      maskLayer.animateFrame(to: bounds, timing: animationTiming)
+      if !maskLayer.hasFrame(bounds) {
+        maskLayer.animateFrame(to: bounds, timing: animationTiming)
+      }
       maskLayer.animate(
         keyPath: "path",
         timing: animationTiming,